[#24] Fix logic errors in setVelocity function
-remove unnecessary recalculation from speed to pwm
-change speed variable type from int to float
-stop directly accessing rotationDirection and use motorDir instead
-also add Talker.h into Test directory as future idea for communications header

+++ robot/rover/ArmDriverUnit/StepperMotor.h
@@ -31,7 +31,7 @@ class StepperMotor : public RobotMotor {
 
     bool calcNumSteps(float angle); // calculates how many steps to take to get to the desired position, assuming no slipping
 
-    void setVelocity(int motorDir, int motorSpeed);
+    void setVelocity(int motorDir, float motorSpeed);
     void stopRotation(void);
 
     // stuff for open loop control
@@ -82,7 +82,7 @@ void StepperMotor::singleStep(int dir) {
   digitalWriteFast(stepPin, LOW);
 }
 
-void StepperMotor::setVelocity(int motorDir, int motorSpeed) {
+void StepperMotor::setVelocity(int motorDir, float motorSpeed) {
   ;
 }
